[MOV] web: move computeAppsAndMenuItems to menu_helpers
X-original-commit: be602afc7e566ab783bcebaf80b8bdb14b6ebba6
