+++ addons/web/static/src/webclient/menus/menu_providers.js
@@ -2,7 +2,7 @@
 
 import { _lt } from "@web/core/l10n/translation";
 import { registry } from "@web/core/registry";
-import { computeAppsAndMenuItems } from "@web/webclient/menus/menu_service";
+import { computeAppsAndMenuItems } from "@web/webclient/menus/menu_helpers";
 import { fuzzyLookup } from "@web/core/utils/search";
 
 const { Component } = owl;
