+++ addons/web/static/src/webclient/menus/menu_service.js
@@ -6,68 +6,6 @@ import { session } from "@web/session";
 
 const loadMenusUrl = `/web/webclient/load_menus`;
 
-/**
- * Traverses the given menu tree, executes the given callback for each node with
- * the node itself and the list of its ancestors as arguments.
- *
- * @param {Object} tree tree of menus as exported by the menus service
- * @param {Function} cb
- * @param {[Object]} [parents] the ancestors of the tree root, if any
- */
-function traverseMenuTree(tree, cb, parents = []) {
-    cb(tree, parents);
-    tree.childrenTree.forEach((c) => traverseMenuTree(c, cb, parents.concat([tree])));
-}
-
-/**
- * Computes the "apps" and "menuItems" from a given menu tree.
- *
- * @param {Object} menuTree tree of menus as exported by the menus service
- * @returns {Object} with keys "apps" and "menuItems" (HomeMenu props)
- */
-export function computeAppsAndMenuItems(menuTree) {
-    const apps = [];
-    const menuItems = [];
-    traverseMenuTree(menuTree, (menuItem, parents) => {
-        if (!menuItem.id || !menuItem.actionID) {
-            return;
-        }
-        const isApp = menuItem.id === menuItem.appID;
-        const item = {
-            parents: parents
-                .slice(1)
-                .map((p) => p.name)
-                .join(" / "),
-            label: menuItem.name,
-            id: menuItem.id,
-            xmlid: menuItem.xmlid,
-            actionID: menuItem.actionID,
-            appID: menuItem.appID,
-        };
-        if (isApp) {
-            if (menuItem.webIconData) {
-                item.webIconData = menuItem.webIconData;
-            } else {
-                const [iconClass, color, backgroundColor] = (menuItem.webIcon || "").split(",");
-                if (backgroundColor !== undefined) {
-                    // Could split in three parts?
-                    item.webIcon = { iconClass, color, backgroundColor };
-                } else {
-                    item.webIconData = "/web_enterprise/static/img/default_icon_app.png";
-                }
-            }
-        } else {
-            item.menuID = parents[1].id;
-        }
-        if (isApp) {
-            apps.push(item);
-        } else {
-            menuItems.push(item);
-        }
-    });
-    return { apps, menuItems };
-}
-
 function makeFetchLoadMenus() {
     const cacheHashes = session.cache_hashes;
     let loadMenusHash = cacheHashes.load_menus || new Date().getTime().toString();
